+++ src/main/java/io/github/shiruka/shiruka/network/Connection.java
@@ -24,8 +24,6 @@
  */
 package io.github.shiruka.shiruka.network;
 
-import io.github.shiruka.shiruka.network.packet.PacketPriority;
-import io.github.shiruka.shiruka.network.packet.PacketReliability;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelHandlerContext;
